requirements.txt
absl-py==2.4.0
accelerate==1.12.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
bitsandbytes==0.49.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.5
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.37.0.dev0
dill==0.3.8
distro==1.9.0
docstring-parser==0.17.0
dotenv==0.9.9
editdistance==0.8.1
einops==0.8.1
evalscope==1.4.2
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.133.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.3.1
gitdb==4.0.12
gitpython==3.1.45
gradio==5.4.0
gradio-client==1.4.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.3
hydra-core==1.3.2
idna==3.11
importlib-metadata==8.7.0
iopath==0.1.10
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia-rs==0.1.10
langdetect==1.0.9
latex2sympy2-extended==1.11.0
lightning-utilities==0.15.2
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
modelscope==1.34.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nltk==3.9.3
numpy==2.2.6
omegaconf==2.3.0
open-clip-torch==3.2.0
openai==2.9.0
openai-clip==1.0.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.7
oss2==2.19.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.2.dev0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
plotly==5.24.1
portalocker==3.2.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==22.0.0
pycocotools==2.0.10
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.12
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
rank-bm25==0.2.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.3.3
rouge-chinese==1.0.3
rouge-score==0.1.2
rpds-py==0.30.0
ruff==0.15.2
sacrebleu==2.6.0
safehttpx==0.1.7
safetensors==0.7.0
sam-2==1.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.47.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.52.1
supervision==0.27.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tokenizers==0.22.1
tomlkit==0.12.0
torch-fidelity==0.3.0
torchmetrics==1.8.2
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.1.dev0
typer==0.24.1
typer-slim==0.21.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.1
uvicorn==0.41.0
wandb==0.23.1
wcwidth==0.2.14
websockets==12.0
word2number==1.1
xxhash==3.6.0
yarl==1.22.0
zhconv==1.4.3
zipp==3.23.0